refactor router initializer
